v1: AutomatedRetrieval: own racks/loading tray and storage bookkeeping
Move the storage site bookkeeping that the legacy Incubator implemented
into the AutomatedRetrieval capability so it is implemented once instead of
per device.

AutomatedRetrieval now takes `racks` and `loading_tray` at construction
(alongside `backend`) and implements get_num_free_sites,
get_site_by_plate_name, find_smallest_site_for_plate, find_random_site,
fetch_plate_to_loading_tray (by plate name), take_in_plate, and summary.
NoFreeSiteError lives here and is re-exported.

Liconic and Cytomat pass their racks/loading tray into the capability and
no longer define these storage methods themselves; callers use them via
`device.retrieval`.
