+++ pylabrobot/capabilities/automated_retrieval/__init__.py
@@ -1,2 +1,2 @@
-from .automated_retrieval import AutomatedRetrieval
+from .automated_retrieval import AutomatedRetrieval, NoFreeSiteError
 from .backend import AutomatedRetrievalBackend
